multi_agent/settings.py
"""Minimal Django settings for the multi-agent demo."""

import os
from pathlib import Path

# --- Load .env to use environment variables ---
try:
    from dotenv import load_dotenv
    load_dotenv()
except Exception:
    pass
# ------------------------------------------------------

BASE_DIR = Path(__file__).resolve().parent.parent

# -----------------
# Core settings
# -----------------
SECRET_KEY = os.getenv("SECRET_KEY", "dev-secret-key")
DEBUG = os.getenv("DJANGO_DEBUG", "False").strip().lower() == "true"

# Comma-separated in .env
ALLOWED_HOSTS = [h.strip() for h in os.getenv("ALLOWED_HOSTS", "").split(",") if h.strip()]

# -----------------
# Installed apps
# -----------------
INSTALLED_APPS = [
    "django.contrib.admin",
    "django.contrib.auth",
    "django.contrib.contenttypes",
    "django.contrib.sessions",
    "django.contrib.messages",
    "django.contrib.staticfiles",

    # Local app
    "chat_api",
]

# -----------------
# Middleware
# -----------------
MIDDLEWARE = [
    "django.middleware.security.SecurityMiddleware",
    "django.contrib.sessions.middleware.SessionMiddleware",
    "django.middleware.common.CommonMiddleware",
    "django.middleware.csrf.CsrfViewMiddleware",
    "django.contrib.auth.middleware.AuthenticationMiddleware",
    "django.contrib.messages.middleware.MessageMiddleware",
    "django.middleware.clickjacking.XFrameOptionsMiddleware",
]

ROOT_URLCONF = "multi_agent.urls"

# -----------------
# Templates (serves templates/index.html)
# -----------------
TEMPLATES = [
    {
        "BACKEND": "django.template.backends.django.DjangoTemplates",
        "DIRS": [BASE_DIR / "templates"],
        "APP_DIRS": True,
        "OPTIONS": {
            "context_processors": [
                "django.template.context_processors.debug",
                "django.template.context_processors.request",
                "django.contrib.auth.context_processors.auth",
                "django.contrib.messages.context_processors.messages",
            ],
        },
    },
]

WSGI_APPLICATION = "multi_agent.wsgi.application"

# -----------------
# Database (SQLite)
# -----------------
DATABASES = {
    "default": {
        "ENGINE": "django.db.backends.sqlite3",
        "NAME": BASE_DIR / "db.sqlite3",
    }
}

# -----------------
# Password validation
# -----------------
AUTH_PASSWORD_VALIDATORS = [
    {"NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"},
    {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator"},
    {"NAME": "django.contrib.auth.password_validation.CommonPasswordValidator"},
    {"NAME": "django.contrib.auth.password_validation.NumericPasswordValidator"},
]

# -----------------
# Internationalization
# -----------------
LANGUAGE_CODE = "en-us"
TIME_ZONE = "UTC"
USE_I18N = True
USE_TZ = True

# -----------------
# Static files
# -----------------
STATIC_URL = "static/"

# -----------------
# OpenAI
# -----------------
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")

# -----------------
# Defaults
# -----------------
DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"
